+++ examples/H5Fed/tetmesh_adjacencies.c
@@ -522,7 +522,7 @@ traverse_mesh (
 	Timer* timer = Timer_.new();
 
 	/* open mesh and get number of levels */
-	printf ("    Opening mesh with id %lld\n", mesh_id);
+	printf ("    Opening mesh with id %lld\n", (long long)mesh_id);
 	H5FedOpenTetrahedralMeshByIndex (f, mesh_id, &mesh);
 	h5_size_t num_levels = H5FedGetNumLevels (mesh);
 	printf ("    Number of levels in mesh: %lld\n", (long long)num_levels);
